Initial version supporting versions of Powershell 2+ and .net 2+
[+] Major rewrite of HookLoginForms, ported a load of the window tracking code into C#.

[+] Found a number of bugs tracking windows, now fixed.

[+] Can't find the HWND (which is used to uniquely identify Tabs) in Posh2 the same way its done in 5.
This required locating it via the Accessibility. The object (IWebBrowser2) does not have frames or parentWindow populated so now required to store a reference to IHTMLWindow2 for script execution.

[+] Have to use Window.execScript rather than reflectively call eval on the document.parentWindow. Script is executed but doesn't seem to return results

The changes meant including the ful pinvoke definition for IWebBrowser2 which is not ideal. I can't find a way to have an Add-Type reference the temporary assembly created by the call to ConvertTypeLibToAssembly